requirements.txt
python-binance
TA-Lib
numpy
websocket_client
sqlalchemy
mysql-connector-python
pandas
pymysql
python-dotenv
pytz
asyncio
nest_asyncio